+++ config/webpack.nonmin.js
@@ -4,25 +4,16 @@ const commonConfig = require('./webpack.common.js')
 const helpers = require('./helpers')
 
 // See https://webpack.js.org/guides/development/
-module.exports = webpackMerge(commonConfig, {
+module.exports = (env) => webpackMerge(commonConfig, {
   mode: 'development',
   devtool: '@source-map',
   output: {
-    path: helpers.root('dist/nonmin'),
+    path: env && env.travis === true ? helpers.root('dist') : helpers.root('dist/nonmin'),
     publicPath: '/dist/',
     filename: '[name].js',
-    chunkFilename: '[id].chunk.js',
     libraryTarget: 'umd'
   },
-
   plugins: [
     new CleanWebpackPlugin() // Cleans directory before building
-  ],
-
-  devServer: {
-    contentBase: helpers.root('.'),
-    historyApiFallback: true,
-    // stats: 'minimal',
-    port: 8080
-  }
+  ]
 })