+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - added option `use_default_values` used to specify if default values based on the "type" of the property have to be used
 - when `use_default_values` is false, number and integer fields have undefined value when input is empty
 - when `use_default_values` is false, string fields have undefined value at the beginning. After that user edit the field, empty input is threated as an empty string
+- Fixed using multiple dependencies for an editor. Fix #703
 
 ### 2.3.0-dev